filemgr/api/views/generic_read_model_view.py
from rest_framework import mixins

from api.views.generic_model_view import GenericModelViewSet


class GenericReadModelViewSet(mixins.RetrieveModelMixin,
                       mixins.ListModelMixin,
                       GenericModelViewSet):
    pass
