filemgr/requirements.txt
asgiref==3.6.0
boto3==1.26.92
botocore==1.29.92
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-extensions==3.2.1
django-q==1.3.9
django-redis==5.2.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
pika==1.3.1
pika-stubs==0.1.3
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.7.1
PyYAML==6.0
rawkit==0.6.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
